+++ psynet/trial/static.py
@@ -361,7 +361,7 @@ class StaticTrialMaker(ChainTrialMaker):
 
 
 class StaticNetwork(ChainNetwork):
-    pass
+    nodes_can_spawn = False
 
 
 class StaticNode(ChainNode):